backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
Pillow>=10.2.0
pillow-avif-plugin>=1.4.3
python-dotenv>=1.0.0
